Communalize tests starting with soundcloud as the test subject
LoveIsGrief/videojs-externals#6 Test youtube
