+++ test/karma.conf.js
@@ -24,8 +24,8 @@ module.exports = function (config) {
     ],
 
     exclude: [
-      // 'test/unit/dailymotion.specs.js',
-      'test/unit/soundcloud.specs.js',
+      'test/unit/dailymotion.specs.js',
+      // 'test/unit/soundcloud.specs.js',
     ],
 
     plugins: [
