+++ test/unit/base/generators/NoSourceTestSuiteGenerator.js
@@ -32,8 +32,8 @@ export default class NoSourceTestSuiteGenerator extends TestSuiteGenerator {
    * @private
    */
   _generateCommonTests() {
-    it('should create widget player', this.basicConfiguration.widgetObjectTest);
-    it('should change object sources', this.basicConfiguration.changeSourceTest(this.secondSourceObject));
-    it('should change string sources', this.basicConfiguration.changeSourceTest(this.secondSourceObject.src));
+    it('should create widget player', this._generateWidgetPlayerTest());
+    it('should change object sources', this._generateChangeSourceTest(this.secondSourceObject));
+    it('should change string sources', this._generateChangeSourceTest(this.secondSourceObject.src));
   }
 }
